feat(useDD.ts) : 제네릭 타입으로 컨트롤러 호출하도록 변경
