+++ components/Modal/InputDropdown/Option.tsx
@@ -2,8 +2,8 @@ import clsx from "clsx";
 import { Dispatch, SetStateAction, useState } from "react";
 import styles from "@/components/Modal/Dropdown/Option.module.css";
 import Image from "next/image";
-import { Member } from "@/hooks/useInputDropdownControll";
 import ProfileIcon from "@/components/header/members/ProfileIcon";
+import { Member } from "@/hooks/useDropdownControll";
 
 interface OptionProp {
   value: any;
